requirements.txt
colorama==0.4.6
llama_index==0.11.1
nest_asyncio==1.6.0
openai==1.42.0
python-dotenv==1.0.1
Requests==2.32.3
